Speed up `bsearch_range_table`.
The comparison function used for the binary search is sub-optimal,
because it looks for an `Equal` result first, even though that is the
least common result (happening 0 or 1 times per search).

This commit changes the comparison function to look for `Equals` last.
As a result, the number of tests executed within the comparison function
is reduced:
- On a `Greater` result: 2 tests --> 1 test
- On a `Less` result: 3 tests --> 2 tests
- On an `Equals` result (next most common): 2 tests --> 2 tests